Separate datasources from default.py
* Move datasources out into their own modules
* Add tests and documentation for datasources
